requirements.txt
aiogram==3.*
aiohttp==3.13.3
python-dotenv==1.2.1
magic-filter==1.0.12
aiofiles==25.1.0
sqlalchemy
greenlet
httpx